SEACR_1.1.R update
Change how the dataframe for density plotting is truncated (previously a hard-coded 90% cutoff): a dataframe of list quantile (i.e. line #/max line#) vs. value quantile (i.e. value/max value) is derived, and the threshold is selected by finding the dataframe pair for which the orthogonal distance below the line defined by (0,0);(1,1) is maximized. Also, the density dataframe is now made to be as long as the integer value closest to 1/10 the length of the input dataframe.